% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-map.R
\name{as_named_colors}
\alias{as_named_colors}
\title{Extract named colors from a persistent map}
\usage{
as_named_colors(x)
}
\arguments{
\item{x}{An \code{sc_color_map}.}
}
\value{
A named hexadecimal character vector.
}
\description{
Extract named colors from a persistent map
}
\examples{
as_named_colors(sc_color_map(c("B", "T")))
}
