% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-palettes.R
\name{sc_color_map_plot}
\alias{sc_color_map_plot}
\title{Plot a persistent color map}
\usage{
sc_color_map_plot(x)
}
\arguments{
\item{x}{An \code{sc_color_map}.}
}
\value{
A ggplot object.
}
\description{
Displays each mapped label alongside its color swatch and hexadecimal code.
}
\examples{
cell_map <- sc_color_map(c("B", "T", "NK"))
sc_color_map_plot(cell_map)
}
