% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{sc_palette_names}
\alias{sc_palette_names}
\title{List registered palette IDs}
\usage{
sc_palette_names(type = NULL, use = NULL, source = NULL, status = NULL)
}
\arguments{
\item{type}{Optional palette type.}

\item{use}{Optional intended-use string.}

\item{source}{Optional source.}

\item{status}{Optional status.}
}
\value{
A character vector of palette IDs.
}
\description{
List registered palette IDs
}
\examples{
sc_palette_names(type = "qualitative")
}
