% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{scChromatic-package}
\alias{scChromatic}
\alias{scChromatic-package}
\title{scChromatic: reproducible color systems for single-cell data}
\description{
scChromatic provides provenance-aware palettes, persistent label-to-color
mappings, ggplot2 scales, deterministic palette extension, and perceptual
diagnostics for single-cell visualizations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://xie186.github.io/scChromatic/}
  \item \url{https://github.com/xie186/scChromatic}
  \item Report bugs at \url{https://github.com/xie186/scChromatic/issues}
}

}
\author{
\strong{Maintainer}: Shaojun Xie \email{xies4@nih.gov} (\href{https://orcid.org/0000-0002-6719-9814}{ORCID}) (Affiliation: FNLCR, Alternate email: xieshaojun0526@gmail.com)

Authors:
\itemize{
  \item Shaojun Xie \email{xies4@nih.gov} (\href{https://orcid.org/0000-0002-6719-9814}{ORCID}) (Affiliation: FNLCR, Alternate email: xieshaojun0526@gmail.com)
}

}
\keyword{internal}
