% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-map.R
\name{update_sc_color_map}
\alias{update_sc_color_map}
\title{Update a persistent color map}
\usage{
update_sc_color_map(map, labels, extend = c("generate", "error"))
}
\arguments{
\item{map}{An \code{sc_color_map}.}

\item{labels}{Labels to add; prior assignments are never changed.}

\item{extend}{Generate or error when palette capacity is exceeded.}
}
\value{
An updated \code{sc_color_map}.
}
\description{
Update a persistent color map
}
\examples{
map <- sc_color_map(c("B", "T"))
update_sc_color_map(map, c("B", "T", "NK"))
}
